move wikidata models models to wikidata package
