+++ app/src/main/java/fr/free/nrw/commons/upload/depicts/DepictsPresenter.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import fr.free.nrw.commons.di.CommonsApplicationModule
 import fr.free.nrw.commons.repository.UploadRepository
 import fr.free.nrw.commons.upload.structure.depictions.DepictedItem
-import fr.free.nrw.commons.data.models.wikidata.WikidataDisambiguationItems
+import fr.free.nrw.commons.wikidata.models.WikidataDisambiguationItems
 import io.reactivex.Flowable
 import io.reactivex.Scheduler
 import io.reactivex.disposables.CompositeDisposable
