+++ app/src/main/java/fr/free/nrw/commons/wikidata/models/WikidataProperties.kt
@@ -1,4 +1,4 @@
-package fr.free.nrw.commons.data.models.wikidata
+package fr.free.nrw.commons.wikidata.models
 
 import fr.free.nrw.commons.BuildConfig
